Cleanup RPM building
- add description
- update license name to match fedora guidelines
- add section bdist_rpm:
    - add packager
    - install docs, include also the example config file
    - add build requires
    - add requires
    - add changelog
- since rst docs aren't very useful, we patch setup.py to create html from them
- add myself to AUTHORS and add entry in ChangeLog or pbr will do so.
Note1: we cannot cleanly handle config files and the example is not directly
useable in any case, so better install as doc
Note2: that we unfortunately cannot easily use a version like 0.9.1.RCauth, since
the build tools change that into a 0.9.2.dev1 in any case.
Note3: that python3-json-logger doesn't exist for EPEL8 but does for F35.
Rebuilding works fine.