Change some of the loglevels
It used to log both a lot of useful and less useful information on level info,
meaning the default log level needed to be info. As a result, it would log 4
lines for each check: Lowering the loglevel of some (debug) messages to log.debug
changes that to 1 line just stating the service is UP. Also raise the level of
"status DOWN" to warning.